test(common): Move tests to misc.rs.
